Split out and added stories for `SidebarSearch`
