+++ lib/ui/package.json
@@ -51,6 +51,7 @@
     "react-lifecycles-compat": "^3.0.4",
     "react-modal": "^3.8.1",
     "react-resize-detector": "^3.2.1",
+    "recompose": "^0.30.0",
     "to-camel-case": "^1.0.0",
     "util-deprecate": "^1.0.2"
   },
