+++ lib/ui/src/components/sidebar/SidebarSearch.stories.js
@@ -1,0 +1,18 @@
+import React from 'react';
+import { action } from '@storybook/addon-actions';
+
+import SidebarSearch, { PureSidebarSearch } from './SidebarSearch';
+
+export default {
+  Component: SidebarSearch,
+  title: 'UI|Sidebar/SidebarSearch',
+};
+
+export const simple = () => <SidebarSearch />;
+
+const onSetFocussed = action('onSetFocussed');
+export const focussed = () => <PureSidebarSearch focussed onSetFocussed={onSetFocussed} />;
+
+export const filledIn = () => (
+  <PureSidebarSearch focussed onSetFocussed={onSetFocussed} defaultValue="Searchstring" />
+);
