+++ lib/ui/src/components/sidebar/explorer-components.js
@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import { styled } from '@storybook/theming';
 
@@ -17,23 +17,6 @@ export const A = styled.a({
   textDecoration: 'none',
 });
 A.displayName = 'A';
-
-const FilterField = styled.input(({ theme }) => ({
-  height: 30,
-  boxSizing: 'border-box',
-  width: '100%',
-  background: 'transparent',
-  border: '0 none',
-  color: theme.mainTextColor,
-  padding: theme.layoutMargin,
-  paddingLeft: 0,
-  outline: 0,
-}));
-const FilterForm = styled.form(({ theme, focussed }) => ({
-  borderBottom: focussed ? `1px solid transparent` : '1px solid transparent',
-  borderBottomColor: focussed ? theme.color.secondary : theme.mainBorderColor,
-  outline: 0,
-}));
 
 const UnstyledRouterLink = styled(RouterLink)({
   color: 'inherit',
@@ -56,34 +39,6 @@ Link.propTypes = {
   prefix: PropTypes.string.isRequired,
   children: PropTypes.node.isRequired,
 };
-
-export class Filter extends Component {
-  state = {
-    focussed: false,
-  };
-
-  set = e => {
-    this.setState({
-      focussed: e.type === 'focus',
-    });
-  };
-
-  render() {
-    const { focussed } = this.state;
-    return (
-      <FilterForm autoComplete="off" focussed={focussed}>
-        <FilterField
-          autocomplete="off"
-          id="storybook-explorer-searchfield"
-          onFocus={this.set}
-          onBlur={this.set}
-          {...this.props}
-          placeholder={focussed ? 'Type to search...' : 'Press "/" to search...'}
-        />
-      </FilterForm>
-    );
-  }
-}
 
 export const StyledLeaf = styled.div(
   ({ theme }) => ({
