+++ lib/ui/src/components/sidebar/explorer.js
@@ -3,7 +3,8 @@ import PropTypes from 'prop-types';
 
 import { Explorer } from '@storybook/components';
 
-import { Section, List, Link, Leaf, Head, Title, Filter } from './explorer-components';
+import { Section, List, Link, Leaf, Head, Title } from './explorer-components';
+import SidebarSearch from './SidebarSearch';
 
 const StoriesPanel = React.memo(props => (
   <Explorer
@@ -14,7 +15,7 @@ const StoriesPanel = React.memo(props => (
     Leaf={Leaf}
     Title={Title}
     Section={Section}
-    Filter={Filter}
+    Filter={SidebarSearch}
   />
 ));
 StoriesPanel.propTypes = {
