+++ lib/ui/src/components/sidebar/explorer.stories.js
@@ -4,7 +4,7 @@ import { storiesOf } from '@storybook/react';
 import { Spaced } from '@storybook/components';
 
 import { StoriesExplorer } from './explorer';
-import { Leaf, Head, Title, Filter } from './explorer-components';
+import { Leaf, Head, Title } from './explorer-components';
 
 import { mockDataset } from './treeview.mockdata';
 
@@ -61,12 +61,5 @@ storiesOf('UI|StoriesExplorer/subcomponents', module)
       <Title type="section" mods={['uppercase']}>
         This is a Root title
       </Title>
-    </Spaced>
-  ))
-  .add('Filter', () => (
-    <Spaced outer={1}>
-      <Filter />
-      <hr />
-      <Filter value="with a value" />
     </Spaced>
   ));
